Move ObjectKey into separate file (out of bean factories).
