+++ airbyte-cdk/bulk/toolkits/load-object-storage/src/main/kotlin/io/airbyte/cdk/load/pipline/object_storage/file/FileCompletedOutputPartitioner.kt
@@ -4,7 +4,7 @@
 
 package io.airbyte.cdk.load.pipline.object_storage.file
 
-import io.airbyte.cdk.load.factory.object_storage.ObjectKey
+import io.airbyte.cdk.load.pipline.object_storage.ObjectKey
 import io.airbyte.cdk.load.file.object_storage.RemoteObject
 import io.airbyte.cdk.load.message.StreamKey
 import io.airbyte.cdk.load.pipline.object_storage.ObjectLoaderCompletedUploadPartitioner
